Move add prettier in write phase and add a note for V7 for prettier
