--- generators/entity-server/phases/configuring/index.js
@@ -1,10 +1,0 @@
-function addPrettier() {
-  if (this.configOptions.generateMfeForEntity) {
-    this.addNpmDevDependency('prettier', '2.0.5');
-    this.addNpmScript('prettier', 'prettier --write "ui/**/*.js"');
-  }
-}
-
-module.exports = {
-  addPrettier,
-};
